requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
passlib[bcrypt]
python-dotenv
pymysql
